+++ tests/migrations/2016_09_06_120008_create_pbk_groups_table.php
@@ -1,0 +1,26 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreatePbkGroupsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up()
+    {
+        Schema::create('pbk_groups', function (Blueprint $table) {
+            $table->increments('ID');
+            $table->string('Name'); //->text('Name');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down()
+    {
+        Schema::drop('pbk_groups');
+    }
+}