+++ CHANGELOG.md
@@ -5,3 +5,13 @@ All notable changes to `gammu` will be documented in this file.
 ## 0.0.1 - 2016-09-02
 
 - Pre-release
+
+## 0.0.2 - 2016-09-07
+
+- Refactor by splitting drivers
+- Add tests
+- Add signature
+
+## 0.0.3 - 2016-09-11
+- Add Callback function for api
+- Remove deadcode